Allow for a table level `default` for removing default constraints with generated names
This method won't actually create a constraint, but it is used for removing constraints
with auto-generated names.

For instance, if you used the default name in
```
schema.create( "my_table", function( table ) {
    table.boolean( "trueOrFalse" ).default( 0 );
} );

```
it may not be obvious that the constraint name is `df_my_table_trueOrFalse`.

This method let's you stay oblivious to the auto-generated name when dropping the constraint.

```
schema.alter( "my_table", function( table ) {
    table.dropConstraint( table.default( "trueOrFalse" ) );
    table.dropColumn( "trueOrFalse" );
} );
```