fix(QueryUtils): Improve numeric sqltype detection
Previously, numeric detection for `inferSQLType` was done
using the `isNumeric` function.  This would return true
for strings that looked like numbers, which would cause
problems in the database if the column was really a string.
This change looks at the underlying Java type to decide
if a variable is a number.  Unfortunately, this doesn't work
for ACF 2016 since they store all types as strings internally.
For ACF 2016, numbers will be passed to the database as `CF_SQL_VARCHAR`
and will need to be set to `CF_SQL_NUMERIC` explicitly. This is
likely to cause less issues than the other way around.

BREAKING CHANGE: On ACF 2016 numeric types will be inferred to be
`CF_SQL_VARCHAR` instead of `CF_SQL_NUMERIC`. You will need to explicitly
pass a param struct to get `CF_SQL_NUMERIC`.