fix(QueryBuilder): Prefer the parent query over magic methods
If a parent query has been set, and a missing method is
found on the parent query, then call it before trying the
magic `where` methods.  This does not change the final try
to the parent query if no methods match.  That is still attempted
which allows the parent query to try `onMissingMethod` magic.