fix(QueryBuilder): Upgrade cbpaginator to fix maxrows discrepency
cbpaginator would clamp the `maxrows` between 0 and a configured
maximum.  qb, however, would execute the query before this clamping.
This means that the length of the results was different than the
reported `maxrows`.  The new version of `cbpaginator` does not do
any clamping.  Please handle that in your application.  This change
is NOT breaking because qb never used the clamped value.  This fix
is just fixing the discrepency between the results length and the
reported maxrows value in the returned pagination struct.