fix(QueryBuilder): Correctly format RETURNING clauses
Specifically when using Quick, column formatters could return
a fully qualified column.  When using a column in a RETURNING clause,
the column should never be fully qualified.  It is always the result
of a query inserting into or updating a specific table.  This fix
uses only the last item in a dot-delimited list as the column name
after formatting.