feat(QueryBuilder): Automatic where scoping with OR clauses in when callbacks
When calling additional QueryBuilder methods inside a `when` callback, it is
easy to forget that those changes, while in a closure, are not in a where group
or scope (meaning parenthesis around the statements).  This commit automatically
scopes where clauses inside a `when` callback, but only if an OR combinator was
used.  This ignores any additional nested queries inside the `when` callback.
This can be turned off using a new fourth parameter, `withoutScoping`.  A
`withScoping` method is publicly available for downstream libraries like
Quick to utilize this scoping feature in additional ways, like Quick scopes
or additional QuickBuilder methods.

BREAKING CHANGE:  Where statements added inside a `when` callback will be
automatically scoped if the callback adds any statements with an OR
combinator.  This change will not likely break any of your code, but it
is technically a breaking change.