feat(QueryBuilder): Add a new simplePaginate method
`simplePaginate` skips the `count` call to the database for instances
when you don't need the count (for total pages) or doing the count
would be a performance hit.  It sends the `results`, `page`, and
`maxRows` parameters to a new `generateSimpleWithResults` on a
`paginationCollector`.  `cbpaginator`, the pagination library that comes
with qb, is set up for this already.  Custom Pagination Collectors
need to add that method to be able to use `simplePaginate`.  Nothing
needs to be done to continue to use `paginate` (hence the minor release).