feat(CI): Adjust fetch depth for tests and release actions
The previous commit had the fetch depths backwards.  It pulled
all commits for tests where we only needed one and only one commit
for release where we need all (to generate a changelog and correctly
do semantic release).  This commit is being marked as a feature
to correctly release the last minor version.