feat(QueryUtils): Automatically add `scale` to bindings when needed
Some versions of ACF combined with SQL Server will automatically use
a `scale` of 0 when not supplied.  This causes very hard to debug issues
since everything looks right from the query side.  To combat this, qb
will automatically provide a scale when the incoming argument is
1) one of the numeric sql types that are floating point and
2) a scale is not already provided in the binding.  This feature can be
disabled by setting `autoAddScale` to `false` in your `moduleSettings`.