Make column formatter lazy
This changes the column formatter to return a struct with `toString` on
it instead of a simple value.  This is so a column formatter can be
changed between the QueryBuilder instance being created and when the
query is compiled.  The initial use case for this is in Quick where
a `hasManyThrough` relationship sets a table alias.  We want all columns
to use that table alias whether they are set before or after setting
the alias.

This does complicate the internals of qb a lot.  While all the tests
do pass, many tests needed to be updated to resolve these values to
strings in order to make the assertions pass.  This might even be
a major release since anyone comparing the values stored in `columns`
or `wheres` or any other value that was previously a string _may_ now
need to resolve the value first. These internals were never part of the
public API, but even libraries like Quick sometimes dip in to it, so
I'd rather be on the safe side and mark this change as breaking.