fix(QueryBuilder): Default columns to empty instead of `*`
In order to preserve the circumstance where someone explicitly
asks for `SELECT *`, we will default to empty columns to start.
When compiling the query, if there are no columns, we will still
`SELECT *` as a default.  The bug fixed here is that if someone
`qb.select( "*" ).addSelectRaw( "ROW_NUMBER() AS rowNumber" )` will
not remove the `SELECT *`.

BREAKING CHANGE: Columns now start as an empty array.  The implicit
`SELECT *` does not get set until compiling the query.  This allows
explicitly adding `.select( "*" )` that is not erased by a
subsequent `addSelect` call.