fix(BaseGrammar): Detect column aliases in raw statements
This is used in conjunction with deriving column names for
`insertUsing` statements.  With this change, the alias
for raw statements can also be derived.  Raw statements
without an alias will not be processed and will result in
an invalid query. In those cases, either an alias can be
added or the `columns` can be explicitly defined.