feat(QueryUtils): Auto derive the correct numeric sql types
In previous versions, qb would use `CF_SQL_NUMERIC` for all
numeric sql types.  This can cause performance issues as
some JDBC drivers interpret `CF_SQL_NUMERIC` as `float` and
then the database has to convert it from there.  This change
introduces some optional settings for `QueryUtils` to solve
this problem: `autoDeriveNumericType`, `integerSqlType`, and
`decimalSqlType`.  When `autoDeriveNumericType` is set to
true, the numeric value will be checked against a regex and
either the `integerSqlType` or the `decimalSqlType` will be
used.  This is an opt-in feature that will likely become
the default in the next major version of qb.