fix(QueryBuilder): Correctly apply native returntypes after `newQuery` and `withReturnFormat`
This would come up often because `paginate` calls do a `count` query
which uses `withReturnFormat`. The `count` couldn't execute because
the `returntype` needed to be cleared when doing a query.  Then we needed
to make sure that the `returntype` was correctly applied after calling
`newQuery` since at that point the `returnFormat` isn't `"array"` but
an identity closure.