feat(QueryBuilder): Add ability to inline bindings when calling `toSQL` and `dump`
This is a highly useful but also potentially dangerous ability - when
generating the SQL for a QueryBuilder you have been able to show the
bindings inline as struct bindings.  If you wanted to then run this in
a SQL client, you had to clean up and format the values.  Now, by passing
`showBindings = "inline"`, qb will attempt to generate a copy and paste
runnable SQL string.  There may be some issues with the type conversion
(especially values like blobs).  Please feel free to open a PR or issue
with improvements.  The danger here is 1) please verify that the generated
string is correct before running it and 2) don't run this string directly
in your application because it is not as safe as running it with a
prepared statement and bindings.