Refactor to polymorphic approach for different join types
We now look for a  "compile#type#join" function to compile each join
in the grammar.  This lets each grammar either support a custom join
implementation or use the built-in implementation without the need
for a `switch` or bunch of `if` statements.  This also paves the way to
supporting `LATERAL` joins for `MySQL` and `Postgres`, which is the
equivalent for those grammars.