feat(QueryBuilder): Support alias renaming using `withAlias`
The `withAlias` function will attempt to rename any currently aliased
columns to the new alias in `columns`, `wheres`, `joins`, `groupBys`,
and `orders`.  This means if you currently have an aliased table like
`users AS "u"` and an aliased column like `"u.name"`, when you call
`qb.withAlias( "u2" )` not only will the table be aliased correctly, but
the column will be aliased correctly as well: `SELECT "u2"."name" FROM "users" AS "u2"`.
Any unaliased columns will be left as-is.