feat(QueryUtils): Auto boolean casting
Grammars will be able to influence the `cfsqltype` and value when passing
in a literal boolean value as a binding.  Postgres and SQLite have
boolean support, so they will keep the literal boolean value and use a
`cfsqltype` of `CF_SQL_OTHER`. SQL Server uses `CF_SQL_BIT`, Oracle users
`CF_SQL_NUMERIC`, and MySQL uses `CF_SQL_TINYINT` — all of these will
convert literal boolean values to either 1 or 0.  This behavior is skipped
when providing a custom `cfsqltype`. Custom grammars can implement the
`getBooleanSqlType` and `convertBooleanValue` methods to customize this
behavior.

BREAKING CHANGE: Previously, literal boolean values would be converted
and treated as `CF_SQL_VARCHAR`. Now they will be converted depending
on the grammar.  Additionally, attempting to change the grammar with
any bindings currently configured will throw an exception.  This is because
the bindings are converted via the grammar when added to the builder
and cannot be changed retroactively when setting a new grammar. Set
the grammar first before configuring the query to avoid this exception.