fix(QueryBuilder): Fix for renaming aliases in subselects
This took a lot of refactoring to store the subselect as a builder
instead of as a raw string.

BREAKING CHANGE: `columns` are now stored as structs, similar to `where`
clauses. Any custom Grammar will need to store `columns` as structs
with at least a `type` column and a `value` column.